requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
alembic==1.8.0
    # via -r requirements.in
anyio==3.6.1
    # via starlette
asgiref==3.5.2
    # via uvicorn
click==8.1.3
    # via uvicorn
colorama==0.4.4
    # via click
fastapi==0.78.0
    # via -r requirements.in
greenlet==1.1.2
    # via sqlalchemy
h11==0.13.0
    # via uvicorn
idna==3.3
    # via anyio
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via mako
mysqlclient==2.1.0
    # via -r requirements.in
pydantic==1.9.1
    # via fastapi
python-dotenv==0.20.0
    # via -r requirements.in
sniffio==1.2.0
    # via anyio
sqlalchemy==1.4.37
    # via
    #   -r requirements.in
    #   alembic
starlette==0.19.1
    # via fastapi
typing-extensions==4.2.0
    # via pydantic
uvicorn==0.17.6
    # via -r requirements.in
